+++ src/components/ui/helper/index.js
@@ -1,0 +1,3 @@
+export { default as TagsAccordionPopover } from "./TagsAccordionPopover";
+export { default as TagsInfoButton } from "./TagsInfoButton";
+export { default as TagsInfoContent } from "./TagsInfoContent";